change app name to freq-split-enhance
