+++ frontend/android/app/src/main/kotlin/com/freqsplit/enhance/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.audioench
+package com.freqsplit.enhance
 
 import io.flutter.embedding.android.FlutterActivity
